OpModeVars --> SharedVars and AutonVars
