+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/control/gainmatrices/RGB.kt
@@ -1,6 +1,6 @@
 package org.firstinspires.ftc.teamcode.control.gainmatrices
 
-import org.firstinspires.ftc.teamcode.opmodes.MainAuton.mTelemetry
+import org.firstinspires.ftc.teamcode.opmodes.SharedVars.mTelemetry
 import kotlin.math.max
 import kotlin.math.min
