+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/control/vision/detectors/AprilTagDetector.java
@@ -1,6 +1,6 @@
 package org.firstinspires.ftc.teamcode.control.vision.detectors;
 
-import static org.firstinspires.ftc.teamcode.opmodes.MainAuton.mTelemetry;
+import static org.firstinspires.ftc.teamcode.opmodes.SharedVars.mTelemetry;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.qualcomm.robotcore.hardware.HardwareMap;
