+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/AutonVars.java
@@ -1,0 +1,17 @@
+package org.firstinspires.ftc.teamcode.opmodes;
+
+import com.acmerobotics.dashboard.config.Config;
+
+@Config
+public final class AutonVars {
+
+    public static double
+            SIZE_WINDOW = 720,
+            LENGTH_ROBOT = 17.3984665354,
+            WIDTH_ROBOT = 16.4220472441,
+            SIZE_HALF_FIELD = 70.5,
+            SIZE_TILE = 23.625,
+            X_START_LEFT = SIZE_TILE * -1.5,
+            X_START_RIGHT = SIZE_TILE * 0.5,
+            Y_START = -SIZE_HALF_FIELD + LENGTH_ROBOT * 0.5;
+}
