+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/SharedVars.java
@@ -1,0 +1,43 @@
+package org.firstinspires.ftc.teamcode.opmodes;
+
+import static org.firstinspires.ftc.teamcode.opmodes.AutonVars.X_START_RIGHT;
+import static org.firstinspires.ftc.teamcode.opmodes.AutonVars.Y_START;
+import static java.lang.Math.PI;
+
+import com.acmerobotics.dashboard.config.Config;
+import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
+import com.acmerobotics.roadrunner.Pose2d;
+import com.arcrobotics.ftclib.gamepad.GamepadEx;
+import com.arcrobotics.ftclib.gamepad.GamepadKeys;
+
+@Config
+public final class SharedVars {
+
+    public static final double
+            REVERSE = PI,
+            LEFT = REVERSE,
+            FORWARD = 1.5707963267948966,
+            RIGHT = 0,
+            BACKWARD = -1.5707963267948966;
+
+    // Declare objects:
+    public static GamepadEx gamepadEx1, gamepadEx2;
+
+    public static boolean keyPressed(int gamepad, GamepadKeys.Button button) {
+        return (gamepad == 2 ? gamepadEx2 : gamepadEx1).wasJustPressed(button);
+    }
+
+    public static MultipleTelemetry mTelemetry;
+
+    static Pose2d robotPose = new Pose2d(X_START_RIGHT, Y_START, FORWARD);
+
+    static boolean isRed = true;
+
+    public static int loopMod(int a, int b) {
+        return (int) loopMod(a,(double) b);
+    }
+
+    public static double loopMod(double a, double b) {
+        return (a % b + b) % b;
+    }
+}
