+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/mechanismtests/TestTeamPropDetector.java
@@ -22,7 +22,7 @@
 package org.firstinspires.ftc.teamcode.opmodes.mechanismtests;
 
 
-import static org.firstinspires.ftc.teamcode.opmodes.MainAuton.mTelemetry;
+import static org.firstinspires.ftc.teamcode.opmodes.SharedVars.mTelemetry;
 
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
