+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/drivetrains/AutoTurner.java
@@ -1,7 +1,7 @@
 package org.firstinspires.ftc.teamcode.subsystems.drivetrains;
 
 import static org.firstinspires.ftc.robotcore.external.navigation.AngleUnit.normalizeRadians;
-import static org.firstinspires.ftc.teamcode.opmodes.MainAuton.mTelemetry;
+import static org.firstinspires.ftc.teamcode.opmodes.SharedVars.mTelemetry;
 import static java.lang.Math.toDegrees;
 
 import com.acmerobotics.dashboard.config.Config;
